server/requirements.txt
fastapi[standard]
python-multipart
langchain 
langchain-google-genai
python-dotenv
pydantic
pypdf2
TTS
pydub